requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via
    #   -r requirements.in
    #   django
attrs==21.4.0
    # via zeep
backports-zoneinfo==0.2.1
    # via django
cached-property==1.5.2
    # via zeep
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via pip-tools
datetime==4.3
    # via -r requirements.in
django==4.0.2
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
idna==3.3
    # via requests
isodate==0.6.1
    # via zeep
lxml==4.7.1
    # via zeep
pep517==0.12.0
    # via pip-tools
pip-tools==6.5.0
    # via -r requirements.in
platformdirs==2.4.1
    # via zeep
python-dotenv==0.19.2
    # via -r requirements.in
pytz==2021.3
    # via
    #   datetime
    #   djangorestframework
    #   zeep
requests==2.27.1
    # via
    #   requests-file
    #   requests-toolbelt
    #   zeep
requests-file==1.5.1
    # via zeep
requests-toolbelt==0.9.1
    # via zeep
six==1.16.0
    # via
    #   isodate
    #   requests-file
sqlparse==0.4.2
    # via django
tomli==2.0.0
    # via pep517
urllib3==1.26.8
    # via requests
wheel==0.37.1
    # via pip-tools
zeep==4.1.0
    # via -r requirements.in
zope-interface==5.4.0
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
